Abandoned strategy for improving brc performance
Bedrock Linux 1.0alpha2 "Momo" performs a number of checks to ensure
things are properly set up before chrooting.  These checks were
initially used for testing, and in theory many are not needed.  Many of
the changes in this commit attempts to reduce the time taken for these
checks to improve performance.  While this as resulted in signifinat
improvements in some areas, there are problems with this strategy in
others.  This strategy has been abandoned in favor of another completely
different strategy in which these checks can be safely removed in their
entirety. This commit/branch is mostly kept in case the new strategy has
unforseen issues and there is interest in returning to this one.