Use specifically libattr v2.4.47
libattr v2.4.48 appears to trigger this issue with gcc:
     https://gcc.gnu.org/bugzilla/show_bug.cgi?id=81523
when building statically across gcc 5.x, 6.x, and 7.x.  gcc 4.x is too
far back to request people utilize.

Debian does not appear to have backported any security updates to
v2.4.47.  So as far as I can tell the v2.4.48 bump was not for
security purposes, and thus sticking with v2.4.47 for the time being is
acceptable.