bri: ls_realpath() explicitly check file existance
Prior to busybox 1.29.x, busybox realpath would only output a path only
if the file exists, and thus the explicit check added here was not
necessary.  However, busybox 1.29.x now returns a path even if the file
does not exist to mimic GNU coreutils, and thus we require this extra
check.

https://git.busybox.net/busybox/commit/?id=747162109fb1c891baf6aafa1ca0410bd08d04a4