strat: clarify capabilities error message
The previous error message confused users who interpreted
"/path/to/strat" as a reference to strata root directories such as
"/bedrock/strata/bedrock".

This new message must be kept up-to-date if strat install location
changes, which is not ideal.  However, the obvious alternative of using
argv[0] is more problematic as argv[0] is likely to report a read-only
crossfs file.