brl-update: scan all mirrors
Originally, brl-update's ability to read multiple mirrors was intended
to allow falling back if the first mirror went down.

This is now being extended to allow users to configure multiple mirrors
such that brl-update will give them the latest update in any mirror.
This covers a mirror going down as well as opens up the option of a beta
channel to be registered along side the stable channel.