Fixed TryExec handling in crossfs ini filter
Previously, crossfs grouped TryExec lines with other Exec lines and
prefixed a strat call to them.  For example, it would convert

    TryExec=/usr/bin/vlc

to

    TryExec=/bedrock/bin/strat debian /usr/bin/vlc

This is correct for other Exec lines, but not for TryExec.  The standard
indicates:

https://specifications.freedesktop.org/desktop-entry-spec/desktop-entry-spec-latest.html

    TryExec Path to an executable file on disk used to determine if the
    program is actually installed. If the path is not an absolute path,
    the file is looked up in the $PATH environment variable. If the file
    is not present or if it is not executable, the entry may be ignored
    (not be used in menus, for example).

TryExec lines should be paths to files which are only checked for
existence and execute permissions; they are never themselves executed.
Thus, the above example should instead be translated to:

    TryExec=/bedrock/strata/debian/usr/bin/vlc