Implement the base for init service translation.
There is now a foundation for cross-stratum service translation. For
now, just the most simple translation is implemented. That is, runit
to systemd. Services are created on the fly and stored in memory
cache, served from there when needed and updated if the original
service file changes. As before, these need to be linked to the init
stratum actual service directory to be used, but it is currently
already possible to run the simples runit services with systemd.