Initial OpenRC service generation implementation
OpenRC is a bit tricky.. it's more complex than sv and it seems to be
easier to just use openrc-run directly, however for that to work
softlevel file has to exist and the openrc command has to be run at
least once so it could create /var/run/openrc stuff. Otherwise,
setting a specific PIDFile for systemd and starting openrc-run seems
to work.