crossfs: Work-around Linux FUSE atomic write issue
This Linux commit:

- 4f06dd92b5d0a6f8eec6a34b8d6ef3e1f4ac1e10
- https://lkml.org/lkml/2021/5/10/1253
- "fuse: fix write deadlock"

broke atomicity of PIPE_BUF-sized writes in FUSE.  This appears to be a
purposeful bug-for-bug trade-off:

> This patch fixes it by only locking the partial tail page.
> If there's a partial head page as well, then split that off as a separate
> WRITE request.

This bug-for-bug trade off was introduced as a bug-fix in a number of
kernels, including but possibly not limited to:

- 5.12.3
- 5.11.20
- 5.10.37
- 5.4.118

This broke crossfs configuration consumption.  Work around it by
buffering multiple writes until we see a newline indicating the
conclusion of a command.