Make hijacked grub themes accessible to GRUB
Some distros configure GRUB to search for themes at a local path which
is valid pre-hijack but not available from GRUB's point of view after
the hijack, as GRUB sees the bedrock stratum.  We can resolve this by
making the GRUB themes available in the corresponding local path in the
bedrock stratum.

This solution is only valid so long as the hijacked GRUB remains in use.
If it is overwritten with another distro's bootloader, it is likely to
break.  Resolving that situation remains an open research problem.