make: more aggressively ensure static linkage
`ldd` returns the same thing if something is statically linked as it
does when it is dynamically linked to a library `ldd` can't find.  Thus,
the return value alone is not a reliable indicator of whether or not
something is statically linked.

In addition, check that `ldd` actually prints to stdout the expected
string indicating the binary is not dynamically linked.