brl-fetch: sort brldb packages by version
brl-fetch was initially naively written without considering the
possibility the repository would have multiple instances of a package.

Ideally eventually we'll fully parse the versions and incorporate them
into our dependency calculations.  Instead, as a quick and dirty work
around, we're sorting the list of packages such that the newest shows up
and is grabbed first.  The hope is that that's the version we want.

Previously, we were naively using `sort -n` to do this sort.  However,
that sorts numerically rather than by version and is confused by more
complex version numbers.  Busybox now supports `sort -V`, and so we
should use that instead.

At the time of writing, this is needed to fix `brl fetch centos`.